+++ packages/stream_chat_flutter_core/CHANGELOG.md
@@ -1,3 +1,17 @@
+## Upcoming
+
+- Updated dependencies to resolvable versions.
+
+## 5.3.0
+
+- Updated `stream_chat` dependency to [`5.3.0`](https://pub.dev/packages/stream_chat/changelog).
+
+## 5.2.0
+
+🔄 Changed
+
+- Updated `connectivity_plus` dependency to `^3.0.2`
+
 ## 5.1.0
 
 - Deprecated the `sort` parameter in the `StreamChannelListController` in favor of `channelStateSort`.